テーブル名をapp_supported_versionに変更
