+++ example-app/santoku-app-backend/src/main/java/jp/fintan/mobile/santokuapp/application/service/system/AppUpdatesService.java
@@ -1,22 +1,22 @@
 package jp.fintan.mobile.santokuapp.application.service.system;
 
 import jp.fintan.mobile.santokuapp.domain.model.system.AppType;
-import jp.fintan.mobile.santokuapp.domain.model.system.AppUpdatesNeeds;
+import jp.fintan.mobile.santokuapp.domain.model.system.AppUpdates;
 import jp.fintan.mobile.santokuapp.domain.model.system.AppVersion;
-import jp.fintan.mobile.santokuapp.domain.repository.AppUpdatesRepository;
+import jp.fintan.mobile.santokuapp.domain.repository.AppSupportedVersionRepository;
 import nablarch.core.repository.di.config.externalize.annotation.SystemRepositoryComponent;
 
 @SystemRepositoryComponent
 public class AppUpdatesService {
 
-  private final AppUpdatesRepository appUpdatesRepository;
+  private final AppSupportedVersionRepository appSupportedVersionRepository;
 
-  public AppUpdatesService(AppUpdatesRepository appUpdatesRepository) {
-    this.appUpdatesRepository = appUpdatesRepository;
+  public AppUpdatesService(AppSupportedVersionRepository appSupportedVersionRepository) {
+    this.appSupportedVersionRepository = appSupportedVersionRepository;
   }
 
-  public AppUpdatesNeeds askAppUpdatesNeeds(AppType appType, AppVersion appVersion) {
-    var appUpdates = appUpdatesRepository.findBy(appType);
-    return new AppUpdatesNeeds(appUpdates, appVersion);
+  public AppUpdates askAppUpdates(AppType appType, AppVersion appVersion) {
+    var appSuppotedVersion = appSupportedVersionRepository.findBy(appType);
+    return new AppUpdates(appSuppotedVersion, appVersion);
   }
 }
