+++ example-app/santoku-app-backend/src/main/java/jp/fintan/mobile/santokuapp/domain/repository/AppSupportedVersionRepository.java
@@ -1,0 +1,9 @@
+package jp.fintan.mobile.santokuapp.domain.repository;
+
+import jp.fintan.mobile.santokuapp.domain.model.system.AppType;
+import jp.fintan.mobile.santokuapp.domain.model.system.AppSupportedVersion;
+
+public interface AppSupportedVersionRepository {
+
+  AppSupportedVersion findBy(AppType appType);
+}
