--- example-app/santoku-app-backend/src/main/java/jp/fintan/mobile/santokuapp/domain/repository/AppUpdatesRepository.java
@@ -1,9 +1,0 @@
-package jp.fintan.mobile.santokuapp.domain.repository;
-
-import jp.fintan.mobile.santokuapp.domain.model.system.AppType;
-import jp.fintan.mobile.santokuapp.domain.model.system.AppUpdates;
-
-public interface AppUpdatesRepository {
-
-  AppUpdates findBy(AppType appType);
-}
