+++ example-app/santoku-app-backend/src/main/resources/db/migration/V2_0__create_table.sql
@@ -1,4 +1,4 @@
-CREATE TABLE app_updates
+CREATE TABLE app_supported_version
 (
   app_type VARCHAR(50) NOT NULL,
   supported_version   VARCHAR(255)  NOT NULL,
