+++ example-app/santoku-app-backend/src/main/resources/db/migration/V2_1__insert_data.sql
@@ -1,2 +1,2 @@
-INSERT INTO app_updates VALUES ('ios', '1.0.0');
-INSERT INTO app_updates VALUES ('android', '1.0.0');
+INSERT INTO app_supported_version VALUES ('ios', '1.0.0');
+INSERT INTO app_supported_version VALUES ('android', '1.0.0');
